refactor(image): #115 워커 모듈 패키지 구조 평탄화 및 예외 격리
