+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/application/dto/PreValidatedImage.java
@@ -1,0 +1,14 @@
+package com.keepgoing.keepgoing.worker.application.dto;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record PreValidatedImage(
+		UUID publicId,
+		String storageKey,
+		String requestedContentType,
+		String detectedContentType,
+		long fileSize,
+		Instant requestedAt
+) {
+}
