+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/application/port/in/ImageProcessingCommand.java
@@ -1,0 +1,13 @@
+package com.keepgoing.keepgoing.worker.application.port.in;
+
+import java.time.Instant;
+import java.util.UUID;
+
+public record ImageProcessingCommand(
+		UUID publicId,
+		String storageKey,
+		String contentType,
+		long fileSize,
+		Instant requestedAt
+) {
+}
