+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/domain/ImageValidationResult.java
@@ -1,0 +1,8 @@
+package com.keepgoing.keepgoing.worker.domain;
+
+public record ImageValidationResult(
+		boolean valid,
+		String detectedContentType,
+		ImageValidationFailureReason reason
+) {
+}
