--- keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/image/application/dto/PreValidatedImage.java
@@ -1,14 +1,0 @@
-package com.keepgoing.keepgoing.worker.image.application.dto;
-
-import java.time.Instant;
-import java.util.UUID;
-
-public record PreValidatedImage(
-		UUID publicId,
-		String storageKey,
-		String requestedContentType,
-		String detectedContentType,
-		long fileSize,
-		Instant requestedAt
-) {
-}
