--- keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/image/application/port/in/ImageProcessingCommand.java
@@ -1,13 +1,0 @@
-package com.keepgoing.keepgoing.worker.image.application.port.in;
-
-import java.time.Instant;
-import java.util.UUID;
-
-public record ImageProcessingCommand(
-		UUID publicId,
-		String storageKey,
-		String contentType,
-		long fileSize,
-		Instant requestedAt
-) {
-}
