--- keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/image/domain/ImageValidationResult.java
@@ -1,8 +1,0 @@
-package com.keepgoing.keepgoing.worker.image.domain;
-
-public record ImageValidationResult(
-		boolean valid,
-		String detectedContentType,
-		ImageValidationFailureReason reason
-) {
-}
