+++ keepgoing-worker/src/main/java/com/keepgoing/keepgoing/worker/infrastructure/redis/adapter/in/RedisImageProcessingRequestListener.java
@@ -1,7 +1,7 @@
 package com.keepgoing.keepgoing.worker.infrastructure.redis.adapter.in;
 
 import com.keepgoing.keepgoing.common.image.event.ImageProcessingRequestedEvent;
-import com.keepgoing.keepgoing.worker.image.application.port.in.ImageProcessingUseCase;
+import com.keepgoing.keepgoing.worker.application.port.in.ImageProcessingUseCase;
 import com.keepgoing.keepgoing.worker.infrastructure.redis.WorkerRedisStreamProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.redis.connection.stream.MapRecord;
